Added tooltip and handler to update node
